+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Column to public domains list which shows the owner of the domain.
+- Official domains are shown with an "official" topic.
 
 ## [1.2.0] - 2021-07-06